+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Multi log support for multiInsert and multiSave operations
 
+### Fixed
+- Array (OR) filters in `getBy` method
+
 ### Changed
 - `getDb()` as public method